PR: Refactor the CartFragment and Finish the skeleton of the MessageFragment

See merge request u7577606/ga-22s2-comp2100-6442!2

+++ app/src/main/java/com/example/mymarketplaceapp/models/Cart.java
@@ -8,9 +8,9 @@ import java.util.List;
 public class Cart implements Parcelable {
     private User user;
     private List<Item> item;
-    private int totalCost;
+    private double totalCost;
 
-    public Cart(User user, List<Item> item, int totalCost) {
+    public Cart(User user, List<Item> item, double totalCost) {
         this.user = user;
         this.item = item;
         this.totalCost = totalCost;
@@ -40,7 +40,7 @@ public class Cart implements Parcelable {
         return item;
     }
 
-    public int getTotalCost() {
+    public double getTotalCost() {
         return totalCost;
     }
 
@@ -55,6 +55,6 @@ public class Cart implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel parcel, int i) {
-        parcel.writeInt(totalCost);
+        parcel.writeDouble(totalCost);
     }
 }
